Added formik props to Input and TextArea component, moved components to Forms folder
